Refactor Edit BudgetGroup and Loading Budget
Editing the BudgetGroup is now like Everydollar! Created two
separate components for simplicity. Added debounce to allow for
deleting (this is what Everydollar does too). Refactored the
loading to be more like a loading and an updating icon, otherwise
the budget goes away every time you make a change which was bad.