Finish Edit BudgetGroupItem
Finish the budget group item. Found a great way to deselect the
current item by knowing when it's clicked outside of. Probably
should refactor BudgetGroup to be the same. Refactor and clean
up code based on eslint/sonarcloud. Trying to add a useCallback
for the refresh of getting new transactions. Hopefully it works!